fix(sync): canonicalize recorded device paths, split planner helpers, extract executor tests (tr-bau)
- record_synced now canonicalizes device-relative paths to forward
  slashes, making recorded state platform-independent: a Windows host
  stringifies planned PathBufs with backslashes, which made the
  executor's Windows CI fail its fingerprint/path assertion and would
  have made a Windows-written state file mismatch planning output on
  every other host. Regression test added.
- split SyncPlanner::plan_pair_tracks (86 LOC, Codacy limit 50) into
  destination_relative_for and delta_kind_for helpers.
- extracted executor tests into device/sync/executor/tests.rs so
  executor/mod.rs drops from 538 to 259 non-comment LOC.
